feat: use single invoices page for paid and unpaid invoices
